html/syntax/parsing/DOMContentLoaded-defer.html

<!doctype html>
<meta charset=utf-8>
<title>The end: DOMContentLoaded and defer scripts</title>
<link rel=help href="https://html.spec.whatwg.org/multipage/#the-end">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var dcl;
var t = async_test(function() {
  dcl = false;
  document.addEventListener("DOMContentLoaded", function(e) {
    dcl = true;
  });
});
</script>
<script defer>
t.step(function() {
  assert_false(dcl, "DOMContentLoaded should not have fired before executing " +
                    "a defer script");

  setTimeout(t.step_func(function() {
    assert_false(dcl, "DOMContentLoaded should not have fired before " +
                      "executing a task queued from a defer script");
    setTimeout(t.step_func_done(function() {
      assert_true(dcl, "DOMContentLoaded should have fired in a task that " +
                       "was queued after the DOMContentLoaded task was queued");
    }), 0);
  }), 0);
});
</script>
